fix: ナビゲーションとカードのコンポーネント見直し
